# Optimize jwt and jws parsing