- Remove unnecessary type conversions
- Optimize when the input is a byte sequence, *bytes.Buffer, *bytes.Reader or *strings.Reader.
- Add benchmark for `jwt.ParseBytes`

Before
```
goos: linux
goarch: amd64
pkg: github.com/lestrrat-go/jwx/jwt
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkJWTParse-16    	  141610	     24251 ns/op	   16352 B/op	      69 allocs/op
PASS
ok  	github.com/lestrrat-go/jwx/jwt	8.545s
```

After
```
goos: linux
goarch: amd64
pkg: github.com/lestrrat-go/jwx/jwt
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkJWTParse-16    	  175860	     15419 ns/op	    7152 B/op	      63 allocs/op
PASS
ok  	github.com/lestrrat-go/jwx/jwt	7.782s
```

In the real world, io.Reader has some similar to byte sequences.
How about the real world use case of jwx?