@lestrrat 2021-01-16
Nice. 

Let me fool around with this for a bit, as I have a couple of things I want to consider. My main concern is the addition of new API. Would it be possible to only introduce ParseBytes() as a public API? This also has to do with the fact that, when the time comes, I've been meaning to change the various Parse(io.Reader) that I had elsewhere in this library with Parse([]byte) and ParseReader(io.Reader) -- I wanted to keep this change until I bump the minor version sometime.