@lestrrat 2021-01-16
I made #289. Some random observations

* ParseString doesn't need to roll its own loop to find the first character. ParseBytes([]byte(s)) is *slightly* slower, but not significantly so. I think the tradeoff of having less code is better than having more code.
* The switch statement to optimized *bytes.Buffer et al have been refactored into a function
* The test coverage went down, so added more tests
* I also added removal warnings for v1.1.0